+++ py/legacyhalos/legacyhalos.py
@@ -349,6 +349,87 @@ def legacyhalos_cosmology(WMAP=False, Planck=False):
 
     return cosmo
 
+def get_lambdabins(verbose=False):
+    """Fixed bins of richness.
+    
+    nn = 7
+    ll = 10**np.linspace(np.log10(5), np.log10(500), nn)
+    #ll = np.linspace(5, 500, nn)
+    mh = np.log10(lambda2mhalo(ll))
+    for ii in range(nn):
+        print('{:.3f}, {:.3f}'.format(ll[ii], mh[ii]))    
+
+    """
+    # Roughly 13.5, 13.9, 14.2, 14.6, 15, 15.7 Msun
+    #lambdabins = np.array([5.0, 10.0, 20.0, 40.0, 80.0, 250.0])
+
+    # Roughly 13.9, 14.2, 14.6, 15, 15.7 Msun
+    lambdabins = np.array([20.0, 25.0, 30.0, 40.0, 60.0, 100.0])
+    #lambdabins = np.array([10.0, 20.0, 40.0, 80.0, 250.0])
+    #lambdabins = np.array([5, 25, 50, 100, 500])
+    nlbins = len(lambdabins)
+    
+    mhalobins = lambda2mhalo(lambdabins, redshift=0.2)
+
+    if verbose:
+        for ii in range(nlbins - 1):
+            print('Bin {}: lambda={:03d}-{:03d}, Mhalo={:.3f}-{:.3f} Msun'.format(
+                ii, lambdabins[ii].astype('int'), lambdabins[ii+1].astype('int'),
+                mhalobins[ii], mhalobins[ii+1]))
+            
+    return lambdabins
+
+def get_zbins(zmin=0.1, zmax=0.3, dt=0.5, verbose=False):
+    """Establish redshift bins which are equal in lookback time."""
+    import astropy.units as u
+    from astropy.cosmology import z_at_value
+    
+    cosmo = legacyhalos_cosmology()
+    tmin, tmax = cosmo.lookback_time([zmin, zmax])
+    if verbose:
+        print('Cosmic time spanned = {:.3f} Gyr'.format(tmax - tmin))
+    
+    ntbins = np.round((tmax.value - tmin.value) / dt + 1).astype('int')
+    #tbins = np.arange(tmin.value, tmax.value, dt) * u.Gyr
+    tbins = np.linspace(tmin.value, tmax.value, ntbins) * u.Gyr
+    zbins = np.around([z_at_value(cosmo.lookback_time, tt) for tt in tbins], decimals=3)
+    tbins = tbins.value
+    
+    # Now fix the bins:
+    # zbins = np.array([0.05, 0.15, 0.25, 0.35, 0.45, 0.6])
+    #zbins = np.array([0.05, 0.1, 0.15, 0.2, 0.25, 0.3, 0.35])
+    zbins = np.array([0.1, 0.15, 0.2, 0.25, 0.3])
+    #zbins = np.array([0.05, 0.1, 0.15, 0.2, 0.25, 0.3, 0.4, 0.5, 0.6])
+    tbins = cosmo.lookback_time(zbins).value
+    
+    if verbose:
+        for ii in range(ntbins - 1):
+            print('Bin {}: z={:.3f}-{:.3f}, t={:.3f}-{:.3f} Gyr, dt={:.3f} Gyr'.format(
+                ii, zbins[ii], zbins[ii+1], tbins[ii], tbins[ii+1], tbins[ii+1]-tbins[ii]))
+            
+    return zbins
+
+def get_mstarbins(deltam=0.1, satellites=False):
+    """Fixed bins of stellar mass.
+    
+    nn = 7
+    ll = 10**np.linspace(np.log10(5), np.log10(500), nn)
+    #ll = np.linspace(5, 500, nn)
+    mh = np.log10(lambda2mhalo(ll))
+    for ii in range(nn):
+        print('{:.3f}, {:.3f}'.format(ll[ii], mh[ii]))    
+    """
+
+    if satellites:
+        pass # code me
+    else:
+        mstarmin, mstarmax = 9.0, 14.0
+
+    nmstarbins = np.round( (mstarmax - mstarmin) / deltam ).astype('int') + 1
+    mstarbins = np.linspace(mstarmin, mstarmax, nmstarbins)
+    
+    return mstarbins
+
 def lambda2mhalo(richness, redshift=0.3, Saro=False):
     """
     Convert cluster richness, lambda, to halo mass, given various 
@@ -375,10 +456,17 @@ def lambda2mhalo(richness, redshift=0.3, Saro=False):
     
     if Saro:
         pass
+
+    if len(np.atleast_1d(redshift)) == 1:
+        zredshift = np.repeat(redshift, len(richness))
+    else:
+        zredshift = redshift
     
     # Melchior et al. 2017 (default)
     logM0, Flam, Gz, lam0, z0 = 14.371, 1.12, 0.18, 30.0, 0.5
-    M200m = 10**logM0 * (richness / lam0)**Flam * ( (1 + redshift) / (1 + z0) )**Gz
+    M200m = 10**logM0 * (richness / lam0)**Flam * ( (1 + zredshift) / (1 + z0) )**Gz
+
+    #np.atleast_1d(M200m)
 
     # Convert to M200c
     #import pdb ; pdb.set_trace()
@@ -387,11 +475,11 @@ def lambda2mhalo(richness, redshift=0.3, Saro=False):
     #M200c, _, _ = mass_adv.changeMassDefinitionCModel(M200m, redshift, '200m', '200c')
 
     # Assume a constant concentration.
-    M200c = np.zeros_like(np.atleast_1d(M200m))
-    for ii, (mm, zz) in enumerate(zip(np.atleast_1d(M200m), np.atleast_1d(redshift))):
+    M200c = np.zeros_like(M200m)
+    for ii, (mm, zz) in enumerate(zip(M200m, zredshift)):
         mc, _, _ = mass_defs.changeMassDefinition(mm, 3.5, zz, '200m', '200c')
         M200c[ii] = mc
-
+        
     return np.log10(M200c)
 
 def cutout_radius_kpc(redshift, pixscale=None, radius_kpc=RADIUS_CLUSTER_KPC, cosmo=None):
